connection发送使用send，并设置MSG_NOSIGNAL

+++ unit_test/echo_server.cc
@@ -91,8 +91,8 @@ int main(int args, char* argv[])
     }
     int     thread_num  = std::stoi(argv[1]);
     char*   ip          = argv[2];
-    int     port        = std::stoi(argv[3]);
-    
+    short   port        = (short)std::stoi(argv[3]);
+
     EchoClient server{thread_num, ip, port};
